Add copy of GIF to develop into SubthreshGIF
SubthreshGIF will be a version of the vanilla GIF module that can be fitted to experimental data that does not contain spikes. Due to details of the implementation of the GIF module, certain parts of the fitting procedure cannot be run if no spikes were found (e.g., the regression fit of the subthreshold dynamics cannot be performed if the spike-triggered current/threshold change are always zero because this results in a singular X matrix).